requirements.txt
python-telegram-bot
firebase-admin
python-dotenv
